enhanced

+++ backend/app/Http/Controllers/TodoController.php
@@ -2,41 +2,49 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Todo;
 
 class TodoController extends Controller
-{   
-    public function index()
 {
-    return Todo::all();
-}
+    public function getTodo()
+    {
+        return response()->json(todo::all(), 200);
+    }
 
-public function store(Request $request)
-{
-    $todo = Todo::create($request->all());
+    public function getTodoId($id)
+    {
+        $todo = Todo::find($id);
+        if (is_null($todo)) {
+            return response()->json(['message' => 'Todo Content not found'], 404);
+        }
+        return response()->json(Todo::find($id), 200);
+    }
 
-    return response()->json($todo, 201);
-}
+    public function addTodo(Request $request)
+    {
+        $todo = Todo::create($request->all());
+        return response($todo, 201);
+    }
 
-public function show(Todo $todo)
-{
-    return response()->json($todo);
-}
+    public function updateTodo(Request $request, $id)
+    {
+        $todo = Todo::find($id);
+        if (is_null($todo)) {
+            return response()->json(['message' => 'Todo Content not found'], 404);
+        }
+        $todo->update($request->all());
+        return response($todo, 200);
+    }
 
-public function update(Request $request, Todo $todo)
-{
-    $todo->update($request->all());
 
-    return response()->json($todo);
-}
-
-public function destroy(Todo $todo)
-{
-    $todo->delete();
-
-    return response()->json(null, 204);
-}
-
+    public function deleteTodo(Request $request, $id)
+    {
+        $todo = Todo::find($id);
+        if (is_null($todo)) {
+            return response()->json(['message' => 'Todo Content not found'], 404);
+        }
+        $todo->delete();
+        return response()->json(null, 204);
+    }
 }
